requirements.txt
openai==1.61.1
python-telegram-bot==21.10
python-dotenv~=1.0.1
httpx~=0.28.1